+++ rating/internal/controller/rating/controller.go
@@ -22,18 +22,18 @@ func NewController(repo ratingRepository) *Controller {
 	return &Controller{repo}
 }
 
-func (c *Controller) GetAggregatedRating(ctx context.Context, recordID model.RecordID, recordType model.RecordType) (float64, error) {
+func (c *Controller) GetAggregatedRating(ctx context.Context, recordID model.RecordID, recordType model.RecordType) (float32, error) {
 	ratings, err := c.repo.Get(ctx, recordID, recordType)
 	if err != nil && errors.Is(err, repository.ErrNotFound) {
 		return 0, ErrNotFound
 	} else if err != nil {
 		return 0, err
 	}
-	sum := float64(0)
+	sum := float32(0)
 	for _, r := range ratings {
-		sum += float64(r.Value)
+		sum += float32(r.Value)
 	}
-	return sum / float64(len(ratings)), nil
+	return sum / float32(len(ratings)), nil
 }
 
 func (c *Controller) PutRating(ctx context.Context, recordID model.RecordID, recordType model.RecordType, rating *model.Rating) error {
